Рефакторинг кода перед групповым ТЗ: - разбить пакеты на несколько подпакетов; - убрать излишнюю таблицу film_mpa_rating; - добавить id рейтинга в таблицу films;

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.model.MPARating;
-import ru.yandex.practicum.filmorate.service.FilmService;
+import ru.yandex.practicum.filmorate.service.film.FilmService;
 
 import java.util.Collection;
 
@@ -72,7 +72,7 @@ public class FilmController {
     }
 
     @GetMapping("/mpa/{id}")
-    public MPARating getMpaRatingById(@PathVariable long id) {
+    public MPARating getMpaRatingById(@PathVariable int id) {
         log.info("==> GET /mpa/" + id);
         MPARating mpaRating = filmService.getMPARatingById(id);
         log.info("<== GET /mpa/" + id
